% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{check_rate_limit}
\alias{check_rate_limit}
\title{Check Rate Limit Status and Sleep If Need Be}
\usage{
check_rate_limit(rsrc, threshold = 2)
}
\arguments{
\item{rsrc}{A resource to check for rate limiting}

\item{threshold}{The remaining API calls below which to force a System Sleep.}
}
\value{
No object returned
}
\description{
Check Rate Limit Status and Sleep If Need Be
}
