% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{check_rate_limit_header}
\alias{check_rate_limit_header}
\title{Check Rate Limit Status and Sleep If Need Be - this is an updated function for API v2.0}
\usage{
check_rate_limit_header(res, threshold = 2)
}
\arguments{
\item{res}{The result of an httr call to the twitter API.}

\item{threshold}{The remaining API calls below which to force a System Sleep.}
}
\value{
binary indicator of whether the API calls had to wait
}
\description{
Check Rate Limit Status and Sleep If Need Be - this is an updated function for API v2.0
}
