% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{get_followers}
\alias{get_followers}
\title{Helper function for get_user_followers}
\usage{
get_followers(user.id, next_token)
}
\arguments{
\item{user.id}{A Twitter user_id}

\item{next_token}{A pagination token from a previous API call.}
}
\value{
A JSON object returned by the Twitter API
}
\description{
This gets called with pagination by get_user_followers
}
