% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{get_user_by_id}
\alias{get_user_by_id}
\title{Get details about a specified user by user_id
see https://developer.twitter.com/en/docs/twitter-api/users/lookup/api-reference/get-users-id}
\usage{
get_user_by_id(user.id)
}
\arguments{
\item{user.id}{A Twitter user_id}
}
\value{
A tibble
}
\description{
Get details about a specified user by user_id
see https://developer.twitter.com/en/docs/twitter-api/users/lookup/api-reference/get-users-id
}
