% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{get_user_by_username}
\alias{get_user_by_username}
\title{Get details about a specified user by username
see https://developer.twitter.com/en/docs/twitter-api/users/lookup/api-reference/get-users-id}
\usage{
get_user_by_username(username)
}
\arguments{
\item{username}{A Twitter username}
}
\value{
A tibble
}
\description{
Get details about a specified user by username
see https://developer.twitter.com/en/docs/twitter-api/users/lookup/api-reference/get-users-id
}
