% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{get_user_followers}
\alias{get_user_followers}
\title{Get list of accounts following a user}
\usage{
get_user_followers(user.id, max.pages = 100)
}
\arguments{
\item{user.id}{The Twitter user_id of the account whose following you want to find}

\item{max.pages}{The maximum number of pages of results to query. Each page returns 1000 users}
}
\value{
A dataframe
}
\description{
This calls get_followers itteratively if pagination is needed
}
\examples{
\dontrun{
  get_user_followers("13205222")
}
}
