% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{get_user_timeline}
\alias{get_user_timeline}
\title{Get user timeline tweets}
\usage{
get_user_timeline(user.id, since.id, start.time, end.time)
}
\arguments{
\item{user.id}{A Twitter user_id}

\item{since.id}{A tweet id. Function then retrieves tweets more recent then that id}

\item{start.time}{YYYY-MM-DDTHH:mm:ssZ (ISO 8601/RFC 3339)}

\item{end.time}{YYYY-MM-DDTHH:mm:ssZ (ISO 8601/RFC 3339)}
}
\value{
A dataframe
}
\description{
see https://developer.twitter.com/en/docs/twitter-api/tweets/timelines/api-reference/get-users-id-tweets
}
