% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{json_followers_to_df}
\alias{json_followers_to_df}
\title{Turn JSON list of users into a tidy dataframe}
\usage{
json_followers_to_df(res)
}
\arguments{
\item{res}{A json object returned by the Twitter API with info about a users followers}
}
\value{
A dataframe with 11 columns
}
\description{
Turn JSON list of users into a tidy dataframe
}
