% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{json_tweets_to_df}
\alias{json_tweets_to_df}
\title{Turn JSON list of tweets into a tidy dataframe}
\usage{
json_tweets_to_df(res)
}
\arguments{
\item{res}{A json object returned by the Twitter API with info about tweets}
}
\value{
A dataframe
}
\description{
Turn JSON list of tweets into a tidy dataframe
}
