% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{json_user_to_df}
\alias{json_user_to_df}
\title{Turn JSON object for single users into a tidy dataframe}
\usage{
json_user_to_df(res)
}
\arguments{
\item{res}{A json object returned by the Twitter API with info about a user}
}
\value{
A dataframe with 11 columns
}
\description{
Turn JSON object for single users into a tidy dataframe
}
