% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{set_bearer_token}
\alias{set_bearer_token}
\title{Set the bearer token for API authentication}
\usage{
set_bearer_token(bearer_token)
}
\arguments{
\item{bearer_token}{An API bearer token from a credentialed twitter app}
}
\description{
need to run this first to set the headers variable which is used in API calls
}
