% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{get_timeline}
\alias{get_timeline}
\title{Helper function for get_user_timeline}
\usage{
get_timeline(user.id, since.id, start.time, end.time)
}
\arguments{
\item{user.id}{A Twitter user_id}
}
\value{
A dataframe
}
\description{
Helper function for get_user_timeline
}
